Fixed File/Pipe/Directory Handling
The functions will now foce the APPDATA\APPNAME directory if it does
belive the path is a full one. On windows it will check for ':' as the
2nd character, everywhere else it looks for '/' in the first character.
On windows it unixPerm is converted to full read, write, execute, or
whatever combination thereof. Later I will fix that by learning ACL but
for now it will suffice.