Touch support sort-of complete
The Button list now scrolls the way it is expected and callbacks
attached to the "buttons" no longer need to call the handler. Now there
are 2 new problems, 1 minor (text is greyed out) & 1 major (cannot
select appropriate buttons when not at zero position). I will consult
the developer of IUP to try and understand what I'm missing if I cannot
find it in the documentation.