Bunches of changes
In my effort to lay the ground work for a simple to understand project
structure that would enable easier porting by other people who are able
to test in enviroments I do not have, I have set about re-organising a
bunch of functions and created more objects & another project which will
handle the actual usage of hacklists. For now I will continue to be the
only contributer until I have programmed a structure I am satisfied with
and is hooking into other processes without issue. For those planning to
aid in the adding of codelists please the ArmaxRaw Library and the
RdTxtHacks.c are what you need to use as a guide.